VCUE-1052 One single container build (#7)

+++ products/podman/src/opensuse/build_in_container.sh
@@ -1,8 +1,17 @@
 #!/usr/bin/bash
 
 cd $(dirname $0)
-zypper install -y libseccomp-devel libgpgme-devel conmon libcontainers-common crun iptables netavark nftables slirp4netns go git libselinux-devel libseccomp-devel python3 man make libbtrfs-devel rpmdevtools gzip rpm-build glib2-devel fdupes glib2-devel-static go-go-md2man golang-packaging libapparmor-devel libostree-devel glibc-devel-static systemd systemd-devel
-ln -s /usr/bin/rpmdev-spectool /usr/bin/spectool
+
+# if it is not a texture-build container, update it
+PACKAGES_FILE="build.packages"
+if [ ! -f /etc/texture.build ] && [ -f "${PACKAGES_FILE}" ]
+then
+  PACKAGES=$(cat ${PACKAGES_FILE} | paste -s -d" ")
+  zypper install -y ${PACKAGES}
+
+  # Add link for spectool
+  ln -s /usr/bin/rpmdev-spectool /usr/bin/spectool
+fi
 
 cd ${PRODUCT}
 make rpm
